+++ Cyotek.Windows.Forms.ColorPicker.Demo/ColorPickerDialogDemoForm.Designer.cs
@@ -28,17 +28,62 @@ namespace Cyotek.Windows.Forms.ColorPicker.Demo
     /// </summary>
     private void InitializeComponent()
     {
+      System.Windows.Forms.Label demoLabel;
       System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(ColorPickerDialogDemoForm));
+      Cyotek.Windows.Forms.GroupBox previewGroupBox;
+      this.dialogColorPreviewPanel = new Cyotek.Windows.Forms.ColorPicker.Demo.ColorPreviewBox();
       this.menuStrip = new System.Windows.Forms.MenuStrip();
       this.fileToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.closeToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.browseColorButton = new System.Windows.Forms.Button();
-      this.demoLabel = new System.Windows.Forms.Label();
-      this.dialogColorPreviewPanel = new Cyotek.Windows.Forms.ColorPicker.Demo.ColorPreviewBox();
       this.colorPreviewPanel = new Cyotek.Windows.Forms.ColorPicker.Demo.ColorPreviewBox();
       this.showAlphaChannelCheckBox = new System.Windows.Forms.CheckBox();
+      this.preserveAlphaChannelCheckBox = new System.Windows.Forms.CheckBox();
+      this.showLoadCheckBox = new System.Windows.Forms.CheckBox();
+      this.showSaveCheckBox = new System.Windows.Forms.CheckBox();
+      demoLabel = new System.Windows.Forms.Label();
+      previewGroupBox = new Cyotek.Windows.Forms.GroupBox();
+      previewGroupBox.SuspendLayout();
       this.menuStrip.SuspendLayout();
       this.SuspendLayout();
+      // 
+      // demoLabel
+      // 
+      demoLabel.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
+            | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+      demoLabel.AutoEllipsis = true;
+      demoLabel.BackColor = System.Drawing.SystemColors.Info;
+      demoLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
+      demoLabel.ForeColor = System.Drawing.SystemColors.InfoText;
+      demoLabel.Location = new System.Drawing.Point(345, 27);
+      demoLabel.Name = "demoLabel";
+      demoLabel.Padding = new System.Windows.Forms.Padding(10);
+      demoLabel.Size = new System.Drawing.Size(412, 310);
+      demoLabel.TabIndex = 6;
+      demoLabel.Text = resources.GetString("demoLabel.Text");
+      // 
+      // previewGroupBox
+      // 
+      previewGroupBox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
+            | System.Windows.Forms.AnchorStyles.Left)));
+      previewGroupBox.Controls.Add(this.dialogColorPreviewPanel);
+      previewGroupBox.Location = new System.Drawing.Point(12, 139);
+      previewGroupBox.Name = "previewGroupBox";
+      previewGroupBox.Size = new System.Drawing.Size(200, 195);
+      previewGroupBox.TabIndex = 3;
+      previewGroupBox.TabStop = false;
+      previewGroupBox.Text = "Preview";
+      // 
+      // dialogColorPreviewPanel
+      // 
+      this.dialogColorPreviewPanel.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+      this.dialogColorPreviewPanel.Color = System.Drawing.Color.Empty;
+      this.dialogColorPreviewPanel.Location = new System.Drawing.Point(6, 25);
+      this.dialogColorPreviewPanel.Name = "dialogColorPreviewPanel";
+      this.dialogColorPreviewPanel.Size = new System.Drawing.Size(188, 164);
+      this.dialogColorPreviewPanel.TabIndex = 1;
       // 
       // menuStrip
       // 
@@ -46,7 +91,8 @@ namespace Cyotek.Windows.Forms.ColorPicker.Demo
             this.fileToolStripMenuItem});
       this.menuStrip.Location = new System.Drawing.Point(0, 0);
       this.menuStrip.Name = "menuStrip";
-      this.menuStrip.Size = new System.Drawing.Size(554, 24);
+      this.menuStrip.Padding = new System.Windows.Forms.Padding(7, 2, 0, 2);
+      this.menuStrip.Size = new System.Drawing.Size(769, 24);
       this.menuStrip.TabIndex = 0;
       // 
       // fileToolStripMenuItem
@@ -63,75 +109,90 @@ namespace Cyotek.Windows.Forms.ColorPicker.Demo
       this.closeToolStripMenuItem.ShortcutKeys = ((System.Windows.Forms.Keys)((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.W)));
       this.closeToolStripMenuItem.Size = new System.Drawing.Size(148, 22);
       this.closeToolStripMenuItem.Text = "&Close";
-      this.closeToolStripMenuItem.Click += new System.EventHandler(this.closeToolStripMenuItem_Click);
+      this.closeToolStripMenuItem.Click += new System.EventHandler(this.CloseToolStripMenuItem_Click);
       // 
       // browseColorButton
       // 
-      this.browseColorButton.Location = new System.Drawing.Point(274, 260);
+      this.browseColorButton.Location = new System.Drawing.Point(12, 106);
       this.browseColorButton.Name = "browseColorButton";
-      this.browseColorButton.Size = new System.Drawing.Size(103, 23);
+      this.browseColorButton.Size = new System.Drawing.Size(121, 27);
       this.browseColorButton.TabIndex = 2;
       this.browseColorButton.Text = "&Choose Color...";
       this.browseColorButton.UseVisualStyleBackColor = true;
-      this.browseColorButton.Click += new System.EventHandler(this.browseColorButton_Click);
-      // 
-      // demoLabel
-      // 
-      this.demoLabel.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-      this.demoLabel.AutoEllipsis = true;
-      this.demoLabel.BackColor = System.Drawing.SystemColors.Info;
-      this.demoLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
-      this.demoLabel.ForeColor = System.Drawing.SystemColors.InfoText;
-      this.demoLabel.Location = new System.Drawing.Point(383, 27);
-      this.demoLabel.Name = "demoLabel";
-      this.demoLabel.Padding = new System.Windows.Forms.Padding(9);
-      this.demoLabel.Size = new System.Drawing.Size(159, 256);
-      this.demoLabel.TabIndex = 3;
-      this.demoLabel.Text = resources.GetString("demoLabel.Text");
-      // 
-      // dialogColorPreviewPanel
-      // 
-      this.dialogColorPreviewPanel.Color = System.Drawing.Color.Empty;
-      this.dialogColorPreviewPanel.Location = new System.Drawing.Point(274, 206);
-      this.dialogColorPreviewPanel.Name = "dialogColorPreviewPanel";
-      this.dialogColorPreviewPanel.Size = new System.Drawing.Size(103, 48);
-      this.dialogColorPreviewPanel.TabIndex = 4;
+      this.browseColorButton.Click += new System.EventHandler(this.BrowseColorButton_Click);
       // 
       // colorPreviewPanel
       // 
       this.colorPreviewPanel.Color = System.Drawing.Color.Empty;
       this.colorPreviewPanel.Location = new System.Drawing.Point(12, 27);
       this.colorPreviewPanel.Name = "colorPreviewPanel";
-      this.colorPreviewPanel.Size = new System.Drawing.Size(256, 256);
-      this.colorPreviewPanel.TabIndex = 5;
+      this.colorPreviewPanel.Size = new System.Drawing.Size(200, 73);
+      this.colorPreviewPanel.TabIndex = 1;
       // 
       // showAlphaChannelCheckBox
       // 
       this.showAlphaChannelCheckBox.Checked = true;
       this.showAlphaChannelCheckBox.CheckState = System.Windows.Forms.CheckState.Checked;
-      this.showAlphaChannelCheckBox.Location = new System.Drawing.Point(274, 27);
+      this.showAlphaChannelCheckBox.Location = new System.Drawing.Point(218, 27);
       this.showAlphaChannelCheckBox.Name = "showAlphaChannelCheckBox";
-      this.showAlphaChannelCheckBox.Size = new System.Drawing.Size(103, 37);
-      this.showAlphaChannelCheckBox.TabIndex = 6;
+      this.showAlphaChannelCheckBox.Size = new System.Drawing.Size(121, 36);
+      this.showAlphaChannelCheckBox.TabIndex = 4;
       this.showAlphaChannelCheckBox.Text = "Show &Alpha Channel";
       this.showAlphaChannelCheckBox.UseVisualStyleBackColor = true;
       // 
+      // preserveAlphaChannelCheckBox
+      // 
+      this.preserveAlphaChannelCheckBox.Location = new System.Drawing.Point(217, 69);
+      this.preserveAlphaChannelCheckBox.Name = "preserveAlphaChannelCheckBox";
+      this.preserveAlphaChannelCheckBox.Size = new System.Drawing.Size(122, 36);
+      this.preserveAlphaChannelCheckBox.TabIndex = 5;
+      this.preserveAlphaChannelCheckBox.Text = "&Preserve Alpha Channel";
+      this.preserveAlphaChannelCheckBox.UseVisualStyleBackColor = true;
+      // 
+      // showLoadCheckBox
+      // 
+      this.showLoadCheckBox.AutoSize = true;
+      this.showLoadCheckBox.Checked = true;
+      this.showLoadCheckBox.CheckState = System.Windows.Forms.CheckState.Checked;
+      this.showLoadCheckBox.Location = new System.Drawing.Point(218, 111);
+      this.showLoadCheckBox.Name = "showLoadCheckBox";
+      this.showLoadCheckBox.Size = new System.Drawing.Size(84, 19);
+      this.showLoadCheckBox.TabIndex = 7;
+      this.showLoadCheckBox.Text = "Show &Load";
+      this.showLoadCheckBox.UseVisualStyleBackColor = true;
+      // 
+      // showSaveCheckBox
+      // 
+      this.showSaveCheckBox.AutoSize = true;
+      this.showSaveCheckBox.Checked = true;
+      this.showSaveCheckBox.CheckState = System.Windows.Forms.CheckState.Checked;
+      this.showSaveCheckBox.Location = new System.Drawing.Point(218, 136);
+      this.showSaveCheckBox.Name = "showSaveCheckBox";
+      this.showSaveCheckBox.Size = new System.Drawing.Size(82, 19);
+      this.showSaveCheckBox.TabIndex = 8;
+      this.showSaveCheckBox.Text = "Show &Save";
+      this.showSaveCheckBox.UseVisualStyleBackColor = true;
+      // 
       // ColorPickerDialogDemoForm
       // 
+      this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
       this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-      this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
-      this.ClientSize = new System.Drawing.Size(554, 297);
+      this.ClientSize = new System.Drawing.Size(769, 346);
+      this.Controls.Add(this.showSaveCheckBox);
+      this.Controls.Add(this.showLoadCheckBox);
+      this.Controls.Add(previewGroupBox);
+      this.Controls.Add(this.preserveAlphaChannelCheckBox);
       this.Controls.Add(this.showAlphaChannelCheckBox);
       this.Controls.Add(this.colorPreviewPanel);
-      this.Controls.Add(this.dialogColorPreviewPanel);
-      this.Controls.Add(this.demoLabel);
+      this.Controls.Add(demoLabel);
       this.Controls.Add(this.browseColorButton);
       this.Controls.Add(this.menuStrip);
+      this.Font = new System.Drawing.Font("Segoe UI", 9F);
       this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
       this.MaximizeBox = false;
       this.Name = "ColorPickerDialogDemoForm";
       this.Text = "ColorPickerDialog Form Demonstration";
+      previewGroupBox.ResumeLayout(false);
       this.menuStrip.ResumeLayout(false);
       this.menuStrip.PerformLayout();
       this.ResumeLayout(false);
@@ -145,9 +206,11 @@ namespace Cyotek.Windows.Forms.ColorPicker.Demo
     private System.Windows.Forms.ToolStripMenuItem fileToolStripMenuItem;
     private System.Windows.Forms.ToolStripMenuItem closeToolStripMenuItem;
     private System.Windows.Forms.Button browseColorButton;
-    private System.Windows.Forms.Label demoLabel;
     private ColorPreviewBox dialogColorPreviewPanel;
     private ColorPreviewBox colorPreviewPanel;
     private System.Windows.Forms.CheckBox showAlphaChannelCheckBox;
+    private System.Windows.Forms.CheckBox preserveAlphaChannelCheckBox;
+    private System.Windows.Forms.CheckBox showLoadCheckBox;
+    private System.Windows.Forms.CheckBox showSaveCheckBox;
   }
 }