+++ Cyotek.Windows.Forms.ColorPicker/ColorPickerDialog.cs
@@ -1,10 +1,13 @@
-// Cyotek Color Picker controls library
-// Copyright © 2013-2021 Cyotek Ltd.
+// Cyotek Color Picker Controls Library
 // http://cyotek.com/blog/tag/colorpicker
 
-// Licensed under the MIT License. See license.txt for the full text.
+// Copyright (c) 2013-2021 Cyotek Ltd.
 
-// If you use this code in your applications, donations or attribution are welcome
+// This work is licensed under the MIT License.
+// See LICENSE.TXT for the full text
+
+// Found this code useful?
+// https://www.cyotek.com/contribute
 
 using System;
 using System.ComponentModel;
@@ -16,60 +19,114 @@ using System.Windows.Forms;
 
 namespace Cyotek.Windows.Forms
 {
-
   [DefaultEvent("PreviewColorChanged")]
   [DefaultProperty("Color")]
   public partial class ColorPickerDialog : Form
   {
-    #region Constants
+    private static readonly object _eventCustomColorsLoading = new object();
+
+    private static readonly object _eventCustomColorsSaving = new object();
 
     private static readonly object _eventPreviewColorChanged = new object();
 
-    #endregion
+    private Color _color;
 
-    #region Fields
+    private ColorCollection _customColors;
+
+    private bool _preserveAlphaChannel;
+
+    private bool _showAlphaChannel;
+
+    private bool _showLoad;
+
+    private bool _showSave;
 
     private Brush _textureBrush;
-
-    #endregion
-
-    #region Constructors
 
     public ColorPickerDialog()
     {
       this.InitializeComponent();
-      this.ShowAlphaChannel = true;
-      this.Font = SystemFonts.DialogFont;
+
+      _showAlphaChannel = true;
+      _showLoad = true;
+      _showSave = true;
+      _customColors = new ColorCollection();
+
+      base.Font = SystemFonts.DialogFont;
     }
 
-    #endregion
+    [Category("Action")]
+    public event CancelEventHandler CustomColorsLoading
+    {
+      add => this.Events.AddHandler(_eventCustomColorsLoading, value);
+      remove => this.Events.RemoveHandler(_eventCustomColorsLoading, value);
+    }
 
-    #region Events
+    [Category("Action")]
+    public event EventHandler CustomColorsSaving
+    {
+      add => this.Events.AddHandler(_eventCustomColorsSaving, value);
+      remove => this.Events.RemoveHandler(_eventCustomColorsSaving, value);
+    }
 
     [Category("Property Changed")]
     public event EventHandler PreviewColorChanged
     {
-      add { this.Events.AddHandler(_eventPreviewColorChanged, value); }
-      remove { this.Events.RemoveHandler(_eventPreviewColorChanged, value); }
+      add => this.Events.AddHandler(_eventPreviewColorChanged, value);
+      remove => this.Events.RemoveHandler(_eventPreviewColorChanged, value);
     }
-
-    #endregion
-
-    #region Properties
 
     public Color Color
     {
-      get { return colorEditorManager.Color; }
-      set { colorEditorManager.Color = value; }
+      get => _color;
+      set
+      {
+        if (_color != value)
+        {
+          colorEditorManager.Color = value;
+
+          _color = value;
+        }
+      }
+    }
+
+    public ColorCollection CustomColors
+    {
+      get => _customColors;
+      set => _customColors = value;
     }
 
     [Browsable(false)]
     [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-    public bool ShowAlphaChannel { get; set; }
+    public bool PreserveAlphaChannel
+    {
+      get => _preserveAlphaChannel;
+      set => _preserveAlphaChannel = value;
+    }
 
-    #endregion
+    [Browsable(false)]
+    [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+    public bool ShowAlphaChannel
+    {
+      get => _showAlphaChannel;
+      set => _showAlphaChannel = value;
+    }
 
-    #region Methods
+    [Browsable(false)]
+    [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+    public bool ShowLoad
+    {
+      get => _showLoad;
+      set => _showLoad = value;
+    }
+
+    [Browsable(false)]
+    [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+    public bool ShowSave
+    {
+      get => _showSave;
+      set => _showSave = value;
+    }
 
     /// <summary>
     /// Clean up any resources being used.
@@ -79,10 +136,7 @@ namespace Cyotek.Windows.Forms
     {
       if (disposing)
       {
-        if (components != null)
-        {
-          components.Dispose();
-        }
+        components?.Dispose();
 
         if (_textureBrush != null)
         {
@@ -95,6 +149,50 @@ namespace Cyotek.Windows.Forms
     }
 
     /// <summary>
+    /// Raises the <see cref="CustomColorsLoading" /> event.
+    /// </summary>
+    /// <param name="e">The <see cref="CancelEventArgs" /> instance containing the event data.</param>
+    protected virtual void OnCustomColorsLoading(CancelEventArgs e)
+    {
+      CancelEventHandler handler;
+
+      handler = (CancelEventHandler)this.Events[_eventCustomColorsLoading];
+
+      if (handler != null)
+      {
+        handler.Invoke(this, e);
+      }
+      else
+      {
+        // if no handler is explicitly bound, then
+        // fall back to the built-in behaviour
+        this.LoadCustomColors();
+      }
+    }
+
+    /// <summary>
+    /// Raises the <see cref="CustomColorsSaving" /> event.
+    /// </summary>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected virtual void OnCustomColorsSaving(EventArgs e)
+    {
+      EventHandler handler;
+
+      handler = (EventHandler)this.Events[_eventCustomColorsSaving];
+
+      if (handler != null)
+      {
+        handler.Invoke(this, e);
+      }
+      else
+      {
+        // if no handler is explicitly bound, then
+        // fall back to the built-in behaviour
+        this.SaveCustomColors();
+      }
+    }
+
+    /// <summary>
     /// Raises the <see cref="E:System.Windows.Forms.Form.Load"/> event.
     /// </summary>
     /// <param name="e">An <see cref="T:System.EventArgs"/> that contains the event data. </param>
@@ -102,25 +200,14 @@ namespace Cyotek.Windows.Forms
     {
       base.OnLoad(e);
 
-      loadPaletteButton.Image = ResourceManager.LoadPalette;
-      savePaletteButton.Image = ResourceManager.SavePalette;
+      this.UpdateButtonStates();
+
       screenColorPicker.Image = ResourceManager.ScreenPicker;
 
-      colorEditor.ShowAlphaChannel = this.ShowAlphaChannel;
+      colorEditor.ShowAlphaChannel = _showAlphaChannel;
+      colorEditor.PreserveAlphaChannel = _preserveAlphaChannel;
 
-      if (!this.ShowAlphaChannel)
-      {
-        for (int i = 0; i < colorGrid.Colors.Count; i++)
-        {
-          Color color;
-
-          color = colorGrid.Colors[i];
-          if (color.A != 255)
-          {
-            colorGrid.Colors[i] = Color.FromArgb(255, color);
-          }
-        }
-      }
+      this.CopyCustomColors();
     }
 
     /// <summary>
@@ -136,20 +223,22 @@ namespace Cyotek.Windows.Forms
       handler?.Invoke(this, e);
     }
 
-    private void cancelButton_Click(object sender, EventArgs e)
+    private void CancelButton_Click(object sender, EventArgs e)
     {
       this.DialogResult = DialogResult.Cancel;
       this.Close();
     }
 
-    private void colorEditorManager_ColorChanged(object sender, EventArgs e)
+    private void ColorEditorManager_ColorChanged(object sender, EventArgs e)
     {
+      _color = colorEditorManager.Color;
+
       previewPanel.Invalidate();
 
       this.OnPreviewColorChanged(e);
     }
 
-    private void colorGrid_EditingColor(object sender, EditColorCancelEventArgs e)
+    private void ColorGrid_EditingColor(object sender, EditColorCancelEventArgs e)
     {
       e.Cancel = true;
 
@@ -166,7 +255,49 @@ namespace Cyotek.Windows.Forms
       }
     }
 
-    private void loadPaletteButton_Click(object sender, EventArgs e)
+    private void CopyCustomColors()
+    {
+      colorGrid.Palette = ColorPalette.None;
+
+      if (_customColors == null)
+      {
+        _customColors = new ColorCollection();
+      }
+
+      if (_customColors.Count == 0)
+      {
+        _customColors.AddRange(ColorPalettes.PaintPalette);
+      }
+
+      colorGrid.BeginUpdate();
+
+      colorGrid.Colors.Clear();
+
+      // copy a maximum of 96 colours from the custom plaette
+      for (int i = 0; i < Math.Min(96, _customColors.Count); i++)
+      {
+        Color color;
+
+        color = _customColors[i];
+
+        if (color.A != 255 && !_showAlphaChannel && !_preserveAlphaChannel)
+        {
+          color = Color.FromArgb(255, color);
+        }
+
+        colorGrid.Colors.Add(color);
+      }
+
+      // or if we have less, fill in the blanks
+      while (colorGrid.Colors.Count < 96)
+      {
+        colorGrid.Colors.Add(Color.White);
+      }
+
+      colorGrid.EndUpdate();
+    }
+
+    private void LoadCustomColors()
     {
       using (FileDialog dialog = new OpenFileDialog
       {
@@ -198,19 +329,8 @@ namespace Cyotek.Windows.Forms
 
               if (palette != null)
               {
-                // we can only display 96 colors in the color grid due to it's size, so if there's more, bin them
-                while (palette.Count > 96)
-                {
-                  palette.RemoveAt(palette.Count - 1);
-                }
-
-                // or if we have less, fill in the blanks
-                while (palette.Count < 96)
-                {
-                  palette.Add(Color.White);
-                }
-
-                colorGrid.Colors = palette;
+                _customColors.Clear();
+                _customColors.AddRange(palette);
               }
             }
             else
@@ -226,13 +346,27 @@ namespace Cyotek.Windows.Forms
       }
     }
 
-    private void okButton_Click(object sender, EventArgs e)
+    private void LoadPaletteButton_Click(object sender, EventArgs e)
+    {
+      CancelEventArgs args;
+
+      args = new CancelEventArgs();
+
+      this.OnCustomColorsLoading(args);
+
+      if (!args.Cancel)
+      {
+        this.CopyCustomColors();
+      }
+    }
+
+    private void OkButton_Click(object sender, EventArgs e)
     {
       this.DialogResult = DialogResult.OK;
       this.Close();
     }
 
-    private void previewPanel_Paint(object sender, PaintEventArgs e)
+    private void PreviewPanel_Paint(object sender, PaintEventArgs e)
     {
       Rectangle region;
 
@@ -256,7 +390,7 @@ namespace Cyotek.Windows.Forms
       e.Graphics.DrawRectangle(SystemPens.ControlText, region.Left, region.Top, region.Width - 1, region.Height - 1);
     }
 
-    private void savePaletteButton_Click(object sender, EventArgs e)
+    private void SaveCustomColors()
     {
       using (FileDialog dialog = new SaveFileDialog
       {
@@ -270,6 +404,7 @@ namespace Cyotek.Windows.Forms
           IPaletteSerializer serializer;
 
           serializer = PaletteSerializer.AllSerializers.Where(s => s.CanWrite).ElementAt(dialog.FilterIndex - 1);
+
           if (serializer != null)
           {
             if (!serializer.CanWrite)
@@ -297,6 +432,37 @@ namespace Cyotek.Windows.Forms
       }
     }
 
-    #endregion
+    private void SavePaletteButton_Click(object sender, EventArgs e)
+    {
+      if (_customColors == null)
+      {
+        _customColors = new ColorCollection();
+      }
+
+      _customColors.Clear();
+      _customColors.AddRange(colorGrid.Colors);
+
+      this.OnCustomColorsSaving(e);
+    }
+
+    private void UpdateButtonStates()
+    {
+      loadPaletteButton.Enabled = _showLoad;
+      savePaletteButton.Enabled = _showSave;
+
+      if (_showLoad || _showSave)
+      {
+        loadPaletteButton.Image = ResourceManager.LoadPalette;
+        loadPaletteButton.Visible = true;
+
+        savePaletteButton.Image = ResourceManager.SavePalette;
+        savePaletteButton.Visible = true;
+      }
+      else
+      {
+        loadPaletteButton.Visible = false;
+        savePaletteButton.Visible = false;
+      }
+    }
   }
 }